requirements.txt
asgiref==3.11.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
idna==3.11
paypal-checkout-serversdk==1.0.3
paypalhttp==1.0.1
pycparser==2.23
pyOpenSSL==25.3.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.4
tzdata==2025.3
urllib3==2.6.2
